Rename "logic" dir to lib and make separate tests dir
