+++ src/lib/schedule.ts
